+++ components/market/properties/popups/featureBox.tsx
@@ -1,0 +1,5 @@
+import { IBusinessHour } from "brancy/models/interfaces";
+
+export default function FeaturesBoxPopup(props: { removeMask: () => void }) {
+  return <></>;
+}